Revert "fix warning C4389", it breaks linux build
This reverts commit 6fe4a07cdd486ee7ad228ee6e3aeb8d5daa1d6d1.

+++ src/game/level/level_editor.c
@@ -308,7 +308,7 @@ int level_editor_render(const LevelEditor *level_editor,
         return -1;
     }
 
-    for (int i = 0; i < LAYER_PICKER_N; ++i) {
+    for (size_t i = 0; i < LAYER_PICKER_N; ++i) {
         if (layer_render(
                 level_editor->layers[i],
                 camera,
